Move LoginModalComponent to components folder
