+++ src/app/pages/website/website.component.ts
@@ -1,6 +1,6 @@
 import { Component } from '@angular/core';
 import { ModalController } from '../../components/modal/modal.controller';
-import { LoginModalComponent } from '../login-modal/login-modal.component';
+import { LoginModalComponent } from '../../components/login-modal/login-modal.component';
 
 @Component({
   selector: 'littil-website',
